fix(render): Rerender also when child files changes
Currently, a knitr/rmarkdown report is only rerun if a) dependent
targets change, b) output files in the e.g. main_files directory
(created by knitr), or c) the main file of the report changes. However,
if the text of a child document changes, a rerun is also desired.

Thus, the current commit adds the child files of a report to the return
value and thus, checking these files for changes.